pyproject.toml
[project]
name = "tunisian-data-collection"
version = "1.0.0"
description = "Scalable data collection for Tunisian Arabic emotion detection"
requires-python = ">=3.9"

dependencies = [
    "google-api-python-client>=2.90.0",
    "tweepy>=4.14.0",
    "praw>=7.7.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
]

[project.scripts]
collect-data = "collect:main"

# Use setuptools instead of hatchling for simpler projects
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
